refactor(beam-stream): extract in-memory repository fakes into reusable test-utils modules
Move InMemoryMovieRepository, InMemoryShowRepository, InMemoryFileRepository,
and InMemoryMediaStreamRepository from the private metadata_tests.rs module into
pub mod in_memory blocks within their respective beam-index repository files,
gated by #[cfg(any(test, feature = "test-utils"))].

Add InMemoryLibraryRepository to beam-index/src/repositories/library.rs with
full implementations of count_files and update_scan_progress.

Propagate beam-stream/test-utils feature to beam-index/test-utils so downstream
crates can access all fakes via a single feature flag.

Update metadata_tests.rs to import from the canonical locations instead of
defining local duplicates.
