+++ beam-index/src/repositories/stream.rs
@@ -148,3 +148,50 @@ impl MediaStreamRepository for SqlMediaStreamRepository {
         Ok(models.into_iter().map(MediaStream::from).collect())
     }
 }
+
+/// In-memory implementation for use in tests and test-utils consumers.
+#[cfg(any(test, feature = "test-utils"))]
+pub mod in_memory {
+    use super::*;
+    use std::collections::HashMap;
+    use std::sync::Mutex;
+
+    #[derive(Debug, Default)]
+    pub struct InMemoryMediaStreamRepository {
+        pub streams: Mutex<HashMap<Uuid, Vec<MediaStream>>>,
+    }
+
+    #[async_trait]
+    impl MediaStreamRepository for InMemoryMediaStreamRepository {
+        async fn insert_streams(&self, streams: Vec<CreateMediaStream>) -> Result<u32, DbErr> {
+            let count = streams.len() as u32;
+            for create in streams {
+                let stream = MediaStream {
+                    id: Uuid::new_v4(),
+                    file_id: create.file_id,
+                    index: create.index,
+                    stream_type: create.stream_type,
+                    codec: create.codec,
+                    metadata: create.metadata,
+                };
+                self.streams
+                    .lock()
+                    .unwrap()
+                    .entry(create.file_id)
+                    .or_default()
+                    .push(stream);
+            }
+            Ok(count)
+        }
+
+        async fn find_by_file_id(&self, file_id: Uuid) -> Result<Vec<MediaStream>, DbErr> {
+            Ok(self
+                .streams
+                .lock()
+                .unwrap()
+                .get(&file_id)
+                .cloned()
+                .unwrap_or_default())
+        }
+    }
+}
